frontend_app/requirements.txt
aiofiles==23.2.1
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.5
exceptiongroup==1.2.0
fastapi==0.104.1
filelock==3.13.1
fonttools==4.45.1
fsspec==2023.10.0
h11==0.14.0
idna==3.5
importlib-resources==6.1.1
Jinja2==3.1.2
kiwisolver==1.4.5
lapx==0.5.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
psutil==5.9.6
py-cpuinfo==9.0.0
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
thop==0.1.1.post2209072238
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.217
urllib3==2.1.0
uvicorn==0.24.0.post1
zipp==3.17.0